+++ src/main/java/com/alimundo/moduloalmacen/RequisicionesSalidas.java
@@ -135,7 +135,7 @@ public class RequisicionesSalidas extends javax.swing.JDialog{
             String categoria;
             String medida;
             String almacen;
-            int cant;
+            double cant;
             String dpto;
             String fecha;
             seleccionrow = tablerequisiciones.getSelectedRow();
@@ -152,7 +152,7 @@ public class RequisicionesSalidas extends javax.swing.JDialog{
                     categoria = rs.getString(3);
                     medida = rs.getString(4);
                     almacen = rs.getString(5);
-                    cant = rs.getInt(6);
+                    cant = rs.getDouble(6);
                     movimientos.salidas.modelosalida.addRow(new Object[]{codigo, nommat, categoria, medida,almacen,String.valueOf(cant)});
                 }
                 ps = con.EstablecerConexion().prepareStatement("EXEC spu_retornafechadptorequisiciones ?");
